all issue-related functions in the same file
