+++ lib/frontend/src/actions/index.js
@@ -55,12 +55,6 @@ export const unsetLoading = target => ({
   target,
 })
 
-export const changeIssue = (newIssue, issueId) => ({
-  type: 'CHANGE_ISSUE',
-  newIssue,
-  issueId,
-})
-
 export const addRepositoryTicket = newTicket => ({
   type: 'ADD_REPOSITORY_TICKET',
   newTicket,
@@ -250,123 +244,6 @@ export const fetchRepository = () => async (dispatch, getState) => {
   }
 }
 
-export const updateATicket = (ticket, values) => async dispatch => {
-  let response, json
-  const { comments } = ticket
-  comments.comments = ticket.comments
-  try {
-    response = await fetch(
-      `/api/repos/${ticket.repo_name}/tickets/${ticket.ticket_number}`,
-      {
-        method: 'PUT',
-        body: JSON.stringify(values),
-        credentials: 'same-origin',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      }
-    )
-  } catch (e) {
-    return dispatch(
-      setError('Erreur pendant la mise à jour du ticket', 'issues')
-    )
-  }
-  if (response.status >= 200 && response.status < 300) {
-    try {
-      json = await response.json()
-    } catch (e) {
-      return dispatch(setError('La réponse ne contient pas de json', 'issues'))
-    }
-    json.objects.comments = comments
-    return dispatch(changeIssue(json.objects, ticket.ticket_id))
-  }
-  const jsonMessage = typeof json === 'undefined' ? '' : json.message
-  return dispatch(
-    setError(
-      `${response.status}: ${response.statusText} ${jsonMessage}`,
-      'issues'
-    )
-  )
-}
-
-const findPriorityIndex = (issue, state) => {
-  const { priority } = state.labels.results
-  priority.priority = state.labels.results.priority
-  for (let i = 0; i < priority.length; i++) {
-    const currentLabel = issue.labels.filter(
-      label => label.label_name === priority[i].text
-    )
-    if (currentLabel.length > 0) {
-      return i
-    }
-  }
-  return null
-}
-
-export const updateIssues = actionType => (dispatch, getState) => {
-  const state = getState()
-  for (const index in state.issues.results.issues) {
-    if (state.issues.results.issues.hasOwnProperty(index)) {
-      const issue = state.issues.results.issues[index]
-      if (issue.selected) {
-        switch (actionType) {
-          case 'removeSelectedIssuesPriority':
-            dispatch(removeTopPriority(issue))
-            break
-          case 'incrementSelectedIssuesPriority':
-            dispatch(incrementPriority(issue))
-            break
-          default:
-            dispatch(closeTicket(issue, state))
-        }
-      }
-    }
-  }
-  return dispatch(unsetLoading('issues'))
-}
-
-export const removeTopPriority = issue => (dispatch, getState) => {
-  const state = getState()
-  return dispatch(
-    updateATicket(issue, {
-      labels: issue.labels.filter(
-        label => label.label_name !== state.labels.results.priority[0].text
-      ),
-    })
-  )
-}
-
-export const incrementPriority = issue => (dispatch, getState) => {
-  const state = getState()
-  const { priority } = state.labels.results
-  priority.priority = state.labels.results.priority
-  let currentLabelPriorityIndex = findPriorityIndex(issue, state)
-  if (currentLabelPriorityIndex === null) {
-    currentLabelPriorityIndex = priority.length - 1
-  }
-  if (currentLabelPriorityIndex > 0) {
-    const newLabelList = issue.labels.filter(
-      label => label.label_name !== priority[currentLabelPriorityIndex].text
-    )
-    newLabelList.push({
-      label_name: priority[currentLabelPriorityIndex - 1].text,
-      color: priority[currentLabelPriorityIndex - 1].color,
-    })
-    return dispatch(updateATicket(issue, { labels: newLabelList }))
-  }
-}
-
-export const closeTicket = (issue, state) => dispatch => {
-  dispatch(
-    updateATicket(issue, {
-      state: 'closed',
-      labels: issue.labels.filter(
-        label => label.label_name !== state.labels.results.priority[0].text
-      ),
-    })
-  )
-}
-
 export const createRepositoryLabels = createdLabels => ({
   type: 'CREATE_LABELS',
   created: createdLabels,
@@ -480,24 +357,3 @@ export const deleteLabels = () => async (dispatch, getState) => {
     dispatch(setErrorRepo(keyIndex(allErrors, 1)))
   }
 }
-
-export const toggleIssue = issueId => ({
-  type: 'TOGGLE_ISSUE',
-  issueId,
-})
-
-export const setIssuesSelectness = (issuesId, isSelected) => ({
-  type: 'SET_ISSUES_SELECTNESS',
-  issuesId,
-  isSelected,
-})
-
-export const toggleExpanded = issueId => ({
-  type: 'TOGGLE_EXPANDED',
-  issueId,
-})
-
-export const toggleCommentsExpanded = issueId => ({
-  type: 'TOGGLE_COMMENTS_EXPANDED',
-  issueId,
-})
